FIT: Large opitmisation(RawReader) and type checking(DataBlockWrapper) for FITRaw, 'expired message' mechanism for FT0 DPL workflow (#5858)
* FT0 LUT prepared as Singleton for DigitBlockFT0 usage

* Large optimization for RawReader in FITRaw library

* Use 'expired messages' mechanism for FT0 raw data decoder (taken from ITS/MFT)

* Mechanism for type checking in DataBlockWrapper, within FITRaw lib

* new LUT

* Test executable for FT0(raw2digit)

Test executable for FT0(raw2digit)

* Hotfix for newLUT branch

Hotfix for newLUT branch

* Hotfix(braces added)

Hotfix(braces added)

Hotfix(braces added)

Hotfix(braces added)

+++ DataFormats/Detectors/FIT/FDD/include/DataFormatsFDD/RawEventData.h
@@ -32,8 +32,8 @@ namespace fdd
 struct EventHeader {
   static constexpr size_t PayloadSize = 16;       //should be equal to 10
   static constexpr size_t PayloadPerGBTword = 16; //should be equal to 10
-  static constexpr int MinNelements = 1;
-  static constexpr int MaxNelements = 1;
+  static constexpr size_t MinNelements = 1;
+  static constexpr size_t MaxNelements = 1;
   union {
     uint64_t word[2] = {};
     struct {
@@ -55,8 +55,8 @@ struct EventHeader {
 struct EventData {
   static constexpr size_t PayloadSize = 5;
   static constexpr size_t PayloadPerGBTword = 10;
-  static constexpr int MinNelements = 1; //additional static field
-  static constexpr int MaxNelements = 12;
+  static constexpr size_t MinNelements = 1; //additional static field
+  static constexpr size_t MaxNelements = 12;
   //
   static constexpr int BitFlagPos = 25; // position of first bit flag(numberADC)
 
@@ -103,8 +103,8 @@ struct EventData {
 struct TCMdata {
   static constexpr size_t PayloadSize = 16;       //should be equal to 10
   static constexpr size_t PayloadPerGBTword = 16; //should be equal to 10
-  static constexpr int MinNelements = 1;
-  static constexpr int MaxNelements = 1;
+  static constexpr size_t MinNelements = 1;
+  static constexpr size_t MaxNelements = 1;
   union {
     uint64_t word[2] = {0};
     struct {
